update arcolinux script

+++ arcolinux-setup.sh
@@ -1,15 +1,12 @@
 #!/bin/bash
 
-# Before running check the following link
-# https://arcolinux.com/things-to-do-after-arcolinux-installation/
-
 # Things to pick during installation:
 # Kernel:
 # 	- default
 #	- zen
 #	- amd/intel ucode(depends on the cpu)
 # Drivers:
-#	- nvidia proprietary + nvidia-dkms/intel open-source/AMD proprietary or open-source
+#	- nvidia proprietary + nvidia-dkms/intel open-source/AMD proprietary or open-source(depends on the gpu)
 # Login:
 #	- whatever suits you(currently at Sddm)
 # Desktop:
@@ -80,7 +77,11 @@
 # Partitions:
 # 	- swap to file
 
-read -r -p "Did you check installation script before running? [Y/n] " response
+echo "Before running check the following link"
+echo "https://arcolinux.com/things-to-do-after-arcolinux-installation/"
+echo "While the script is running, you should customize your theme and change default apps if needed."
+
+read -r -p "Have you check installation script before running? [Y/n] " response
 if [[ "$response" =~ ^([nN][eE][sS]|[nN])$ ]]
 then
     echo "Please check it."
@@ -111,14 +112,14 @@ chsh -s /bin/zsh
 git config --global user.name "Dejan Zdravkovic"
 git config --global user.email "bagzi1996@gmail.com"
 
-# mysql
+# mysql (current password for root is empty)
 sudo pacman -S mariadb
 sudo mysql_install_db --user=mysql --basedir=/usr --datadir=/var/lib/mysql
 sudo systemctl enable --now mariadb
 sudo mysql_secure_installation
 sudo mysql -e "ALTER USER 'root'@'localhost' IDENTIFIED BY 'root';"
 
-# postgres (pass: root)
+# postgres (set password to "root")
 sudo pacman -S postgresql
 sudo su - postgres -c "initdb -D '/var/lib/postgres/data'"
 sudo systemctl enable postgresql
@@ -162,6 +163,9 @@ yay gwe
 # rgb config (r: 200, g: 140: b:255)
 yay openrgb
 
-reboot
-
-# After the script is done you should choose your themes and default apps manually.
+# reboot if needed
+read -r -p "Do you want to reboot now? [y/N] " response
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]
+then
+    reboot
+fi
